Fixed bug with modifiers not working. Added UI cursor. Fixed bug with event system not being available by the first update loop of a scene.

+++ Assets/Scripts/Attributes/Health.cs
@@ -109,7 +109,7 @@ namespace RPG.Attributes
 
         public void RestoreState(object state)
         {
-            var healthState = (int)state;
+            var healthState = (float)state;
             _healthPoints.value = healthState;
             
             if (_healthPoints.value == 0)
